fix bug which prevents hmr on scss files
